# src/docs/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Evently API",
    "description": "API for Evently",
    "version": "1.0.0"
  },
  "host": [
    "theodev.myftp.org:89"
  ],
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter your Bearer token in the format **Bearer <token>**"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/user/friends": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Get user friends.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User friends retrieved successfully.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/user/me": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Get all information of the connected user.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User information retrieved successfully.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "401": {
            "description": "Unauthorized. No valid token provided.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 401
                },
                "msg": {
                  "type": "string",
                  "example": "Unauthorized."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/user/": {
      "delete": {
        "tags": [
          "User"
        ],
        "description": "Delete an User.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User information deleted successfully.",
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "example": ""
                },
                "status": {
                  "type": "number",
                  "example": 200
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized. No valid token provided.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 401
                },
                "msg": {
                  "type": "string",
                  "example": "Unauthorized."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      },
      "put": {
        "tags": [
          "User"
        ],
        "description": "Update user details.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "User update data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User updated successfully.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 404
                },
                "msg": {
                  "type": "string",
                  "example": "User not found."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "422": {
            "description": "Unprocessable entity.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 422
                },
                "msg": {
                  "type": "string",
                  "example": "Required fields missing."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/user/register": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Register a new user.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "User registration data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/user/available/:email": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Check if an email is available.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "User login.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "User login data",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "user@example.com"
                },
                "password": {
                  "type": "string",
                  "example": "Password123"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User login successful.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/user/refresh-token/{refreshToken}": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Refresh user token.",
        "parameters": [
          {
            "name": "refreshToken",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Refresh token"
          }
        ],
        "responses": {
          "200": {
            "description": "User token refreshed successfully.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "401": {
            "description": "Unauthorized. No valid token provided.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 401
                },
                "msg": {
                  "type": "string",
                  "example": "Aucun token valide trouvé."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    }
  },
  "definitions": {
    "Category": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 666
        },
        "name": {
          "type": "string",
          "example": "foobar"
        },
        "icon": {
          "type": "string",
          "example": "foobar"
        },
        "color": {
          "type": "string",
          "example": "foobar"
        },
        "default": {
          "type": "boolean",
          "example": true
        },
        "user": {
          "$ref": "#/definitions/User"
        }
      }
    },
...
    "Event": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 666
        },
        "name": {
          "type": "string",
          "example": "foobar"
        },
        "description": {
          "type": "string",
          "example": "foobar"
        },
        "location": {
          "type": "string",
          "example": "foobar"
        },
        "childCreated": {
          "type": "boolean",
          "example": true
        },
        "recurrencePattern": {
          "type": "string",
          "example": "monthly"
        },
        "interval": {
          "type": "number",
          "example": 666
        },
        "maxOccurence": {
          "type": "number",
          "example": 666
        },
        "targetDate": {
          "type": "string",
          "example": "2019-01-01T00:00:00.000Z"
        },
        "inviteToken": {
          "type": "string",
          "example": "foobar"
        },
        "categoryID": {
          "type": "number",
          "example": 666
        },
        "userID": {
          "type": "number",
          "example": 666
        },
        "folderID": {
          "type": "number",
          "example": 666
        },
        "createdAt": {
          "type": "string",
          "example": "2019-01-01T00:00:00.000Z"
        },
        "deletedAt": {
          "type": "string",
          "example": "2019-01-01T00:00:00.000Z"
        },
        "joinedUser": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/JoinedEventEntity"
          }
        }
      }
    },
...
    "JoinedEventEntity": {
      "type": "object",
      "properties": {
        "leaveDate": {
          "type": "string",
          "example": "2019-01-01T00:00:00.000Z"
        },
        "invitationStatus": {
          "type": "string",
          "example": "invited"
        },
        "event": {
          "$ref": "#/definitions/Event"
        },
        "user": {
          "$ref": "#/definitions/User"
        }
      }
    },
...
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 666
        },
        "username": {
          "type": "string",
          "example": "foobar"
        },
        "firstName": {
          "type": "string",
          "example": "foobar"
        },
        "lastName": {
          "type": "string",
          "example": "foobar"
        },
        "email": {
          "type": "string",
          "example": "foobar"
        },
        "sub": {
          "type": "string",
          "example": "foobar"
        },
        "password": {
          "type": "string",
          "example": "foobar"
        },
        "provider": {
          "type": "string",
          "example": "foobar"
        },
        "isDeleted": {
          "type": "boolean",
          "example": true
        },
        "isCompleted": {
          "type": "boolean",
          "example": true
        },
        "createdAt": {
          "type": "string",
          "example": "2019-01-01T00:00:00.000Z"
        },
        "deletedAt": {
          "type": "string",
          "example": "2019-01-01T00:00:00.000Z"
        },
        "categories": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Category"
          }
        },
        "events": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Event"
          }
        },
        "joinedEvents": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/JoinedEventEntity"
          }
        }
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}